third_party/blink/web_tests/platform/mac-mac15-arm64/http/tests/security/webgl-cross-origin-ImageBitmap-blocked-expected.txt

CONSOLE ERROR: Uncaught TypeError: Cannot read properties of null (reading 'createTexture')
WebGL's tex(Sub)Image2D should throw a SecurityError exception when the ImageBitmap is not origin clean.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL Unexpected error: Uncaught TypeError: Cannot read properties of null (reading 'createTexture')
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
